chore: constant, handleError, utils 파일 분리
